requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
conda==4.6.14
cryptography==2.6.1
Django==2.2.3
djangorestframework==3.10.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
meld3==1.0.2
mysqlclient==1.4.2.post1
pipreqs==0.4.9
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.6.8
pytz==2019.1
requests==2.21.0
ruamel-yaml==0.15.46
six==1.12.0
sqlparse==0.3.0
supervisor==4.0.3
urllib3==1.24.1
virtualenv==16.6.2
yarg==0.1.9
